refactored usages of any removed unused util class and fixed imports
