+++ lib/types/index.ts
@@ -1,4 +1,3 @@
 export * from './Messages';
 export * from './Internal.types';
-export * from './SSI.types';
 export * from './SSITypesBuilder';
