+++ lib/utils/index.ts
@@ -1,1 +1,3 @@
 export * from './jsonPathUtils';
+export * from './ObjectUtils';
+export * from './ObjectValidationUtils';
