+++ lib/validation/bundlers/inputDescriptorsV2VB.ts
@@ -1,6 +1,6 @@
 import { ConstraintsV2, FieldV2, HolderSubject, InputDescriptorV2 } from '@sphereon/pex-models';
 
-import { ObjectValidationUtils } from '../../utils/ObjectValidationUtils';
+import { ObjectValidationUtils } from '../../utils';
 import { Validation } from '../core';
 
 import { ConstraintsVB } from './constraintsVB';
